Move booking-related models into a separate file
